# api/openapi.yaml
openapi: 3.0.3
info:
  title: Carpooling VE API
  description: |
    API para la plataforma de carpooling interurbano de Venezuela.
    Corredor inicial: Guatire ↔ Caracas.
    MVP — Mes 3.
  version: 0.1.0
  contact:
    name: Engineer
servers:
  - url: https://api.staging.carpooling.ve
    description: Staging
  - url: http://localhost:3000
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar nuevo usuario (pasajero o conductor)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario creado, OTP enviado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verificar código OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone:
                  type: string
                  example: "+584121234567"
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: OTP verificado, sesión iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Código inválido o expirado
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión con teléfono + OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  example: "+584121234567"
      responses:
        '200':
          description: OTP enviado al teléfono
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Reservar puesto en un viaje
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201':
          description: Reserva creada (pendiente de pago)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
    get:
      tags: [Bookings]
      summary: Listar mis reservas
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled, no_show]
      responses:
        '200':
          description: Lista de reservas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: Detalle de reserva
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Reserva encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /bookings/{id}/cancel:
    post:
      tags: [Bookings]
      summary: Cancelar reserva (aplica política de reembolso)
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Reserva cancelada, reembolso procesado según política
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /bookings/{id}/accept:
    post:
      tags: [Bookings]
      summary: Conductor acepta solicitud de reserva
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Reserva confirmada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /bookings/{id}/reject:
    post:
      tags: [Bookings]
      summary: Conductor rechaza solicitud de reserva
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Reserva rechazada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'

  # ─── PAYMENTS ───
  /payments:
    post:
      tags: [Payments]
      summary: Procesar pago de reserva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      responses:
        '201':
          description: Pago registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/{id}:
    get:
      tags: [Payments]
      summary: Detalle de pago
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Pago encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/{id}/confirm:
    post:
      tags: [Payments]
      summary: Confirmar recepción de transferencia (admin / conciliación)
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Pago confirmado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'

  # ─── REVIEWS ───
  /reviews:
    post:
      tags: [Reviews]
      summary: Calificar al otro participante del viaje
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Calificación registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
  /reviews/user/{userId}:
    get:
      tags: [Reviews]
      summary: Calificaciones recibidas por un usuario
      security: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Lista de calificaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'

  # ─── INCIDENTS ───
  /incidents:
    post:
      tags: [Incidents]
      summary: Reportar incidente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreate'
      responses:
        '201':
          description: Incidente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /incidents/panic:
    post:
      tags: [Incidents]
      summary: Botón de pánico — notifica inmediatamente a operaciones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trip_id, lat, lng]
              properties:
                trip_id:
                  type: string
                  format: uuid
                lat:
                  type: number
                  format: float
                lng:
                  type: number
                  format: float
                message:
                  type: string
      responses:
        '201':
          description: Alerta de pánico enviada a operaciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'

  # ─── CHAT ───
...
  /chat/conversations/{id}/messages:
    get:
      tags: [Chat]
      summary: Mensajes de una conversación
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: before
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Mensajes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
    post:
      tags: [Chat]
      summary: Enviar mensaje
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        '201':
          description: Mensaje enviado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'

  # ─── ADMIN ───
...
  /admin/drivers/{id}/approve:
    post:
      tags: [Admin]
      summary: Aprobar conductor
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Conductor aprobado
  /admin/drivers/{id}/reject:
    post:
      tags: [Admin]
      summary: Rechazar conductor
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Conductor rechazado
...
  /admin/payments/report:
    get:
      tags: [Admin]
      summary: Reporte de comisiones (semanal/mensual)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [weekly, monthly]
            default: weekly
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Reporte de comisiones
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_payments:
                    type: integer
                  total_gross:
                    type: number
                    format: float
                  total_commission:
                    type: number
                    format: float
                  total_driver_payout:
                    type: number
                    format: float
                  by_method:
                    type: object
  /admin/incidents:
    get:
      tags: [Admin]
      summary: Cola de incidentes
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, investigating, resolved]
      responses:
        '200':
          description: Lista de incidentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incident'
  /admin/audit-log:
    get:
      tags: [Admin]
      summary: Audit log de acciones administrativas
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: actor_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Entradas del audit log
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    actor_id:
                      type: string
                    action:
                      type: string
                    target_type:
                      type: string
                    target_id:
                      type: string
                    metadata:
                      type: object
                    created_at:
                      type: string
                      format: date-time
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BookingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PaymentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ConversationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    RegisterRequest:
      type: object
      required: [name, phone, id_doc, role]
      properties:
        name:
          type: string
          example: "Carlos Rodríguez"
        phone:
          type: string
          example: "+584121234567"
        email:
          type: string
          format: email
        id_doc:
          type: string
          description: Cédula de identidad
          example: "V-12345678"
        role:
          type: string
          enum: [passenger, driver]
        selfie_url:
          type: string
          description: URL de la selfie subida
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token
        user:
          $ref: '#/components/schemas/User'
        otp_sent:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        id_doc:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        status:
          type: string
          enum: [active, suspended, blocked]
        rating:
          type: number
          format: float
          description: Calificación promedio
        completed_trips:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trip_id:
          type: string
          format: uuid
        passenger_id:
          type: string
          format: uuid
        seats:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no_show]
        payment_id:
          type: string
          format: uuid
          nullable: true
        cancellation_reason:
          type: string
          nullable: true
        refund_amount:
          type: number
          format: float
          nullable: true
        refund_percentage:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
    BookingCreate:
      type: object
      required: [trip_id, seats]
      properties:
        trip_id:
          type: string
          format: uuid
        seats:
          type: integer
          minimum: 1
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        booking_id:
          type: string
          format: uuid
        amount_usd:
          type: number
          format: float
        amount_bs:
          type: number
          format: float
          nullable: true
        commission_usd:
          type: number
          format: float
          description: 15% del monto
        driver_amount_usd:
          type: number
          format: float
          description: Monto neto al conductor
        method:
          type: string
          enum: [transfer_usd, zelle, pago_movil, cash]
        status:
          type: string
          enum: [pending, confirmed, disputed, refunded]
        reference:
          type: string
          description: Número de referencia de la transferencia
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
    PaymentCreate:
      type: object
      required: [booking_id, method, amount_usd]
      properties:
        booking_id:
          type: string
          format: uuid
        method:
          type: string
          enum: [transfer_usd, zelle, pago_movil, cash]
        amount_usd:
          type: number
          format: float
        amount_bs:
          type: number
          format: float
          nullable: true
        reference:
          type: string
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trip_id:
          type: string
          format: uuid
        from_user_id:
          type: string
          format: uuid
        to_user_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        created_at:
          type: string
          format: date-time
    ReviewCreate:
      type: object
      required: [trip_id, to_user_id, rating]
      properties:
        trip_id:
          type: string
          format: uuid
        to_user_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    Incident:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trip_id:
          type: string
          format: uuid
        reporter_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [panic, late_cancel, wrong_vehicle, accident, payment_fraud, conflict, other]
        description:
          type: string
        status:
          type: string
          enum: [open, investigating, resolved]
        location:
          $ref: '#/components/schemas/GeoPoint'
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
    IncidentCreate:
      type: object
      required: [trip_id, type]
      properties:
        trip_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [late_cancel, wrong_vehicle, accident, payment_fraud, conflict, other]
        description:
          type: string
        lat:
          type: number
        lng:
          type: number
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
    GeoPoint:
      type: object
      properties:
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float